labs/1-development-of-a-multilayer-perceptron-neuroemulator/resolve/simple_emulator/requirements.txt
pytest
fastapi
loguru
uvicorn
opencv-python-headless
python-multipart
numpy
sqlalchemy
psycopg2-binary
PyJWT
python-dotenv
